Refactor lab and user database structure
- Removed old request and user database files, consolidating functionality into a new lab structure.
- Introduced a new lab module with member management, role handling, and permissions.
- Updated request handling to align with the new lab structure, including print requests and tags.
- Enhanced user API key management by separating it into its own module.
- Improved middleware for authentication and logging.
- Adjusted HTTP routes to reflect the new database structure and ensure proper authorization checks.
